@@ -1,6 +1,6 @@
+import 'dart:math' as math;
 
 import 'package:flutter/material.dart';
-import 'dart:math' as math;
 
 Widget _defaultTransitionBuilder(Widget child, Animation<double> animation) =>
     ScaleTransition(
@@ -25,18 +25,18 @@ class IconToggle extends StatefulWidget {
   final Color activeColor;
   final Color inactiveColor;
   final bool value;
-  final ValueChanged<bool> onChanged;
+  final ValueChanged<bool>? onChanged;
   final AnimatedSwitcherTransitionBuilder transitionBuilder;
   final Duration duration;
-  final Duration reverseDuration;
+  final Duration? reverseDuration;
   @override
   _IconToggleState createState() => _IconToggleState();
 }
 
 class _IconToggleState extends State<IconToggle>
     with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-  Animation<double> _position;
+  AnimationController? _controller;
+  Animation<double>? _position;
   bool _cancel = false;
 
   @override
@@ -46,19 +46,21 @@ class _IconToggleState extends State<IconToggle>
         vsync: this,
         duration: Duration(milliseconds: 100),
         reverseDuration: Duration(milliseconds: 50));
-    _position = CurvedAnimation(parent: _controller, curve: Curves.linear);
-    _position.addStatusListener((status) {
+    _position = CurvedAnimation(parent: _controller!, curve: Curves.linear);
+    _position!.addStatusListener((status) {
       if (status == AnimationStatus.dismissed &&
           widget.onChanged != null &&
           _cancel == false) {
-        widget.onChanged(!widget.value);
+        widget.onChanged!(!widget.value);
       }
     });
   }
 
   @override
   void dispose() {
-    _controller?.dispose();
+    if (_controller != null) {
+      _controller!.dispose();
+    }
     super.dispose();
   }
 
@@ -80,7 +82,7 @@ class _IconToggleState extends State<IconToggle>
       child: Padding(
         padding: const EdgeInsets.all(10.0),
         child: _IconToggleable<double>(
-          listenable: _position,
+          listenable: _position!,
           activeColor: widget.activeColor,
           inactiveColor: widget.inactiveColor,
           child: AnimatedSwitcher(
@@ -88,7 +90,9 @@ class _IconToggleState extends State<IconToggle>
             reverseDuration: widget.reverseDuration,
             transitionBuilder: widget.transitionBuilder,
             child: Icon(
-              widget.value ? widget.selectedIconData : widget.unselectedIconData,
+              widget.value
+                  ? widget.selectedIconData
+                  : widget.unselectedIconData,
               color: widget.value ? widget.activeColor : widget.inactiveColor,
               size: 22,
               key: ValueKey<bool>(widget.value),
@@ -102,10 +106,10 @@ class _IconToggleState extends State<IconToggle>
 
 class _IconToggleable<T> extends AnimatedWidget {
   _IconToggleable({
-    Animation<T> listenable,
-    this.activeColor,
-    this.inactiveColor,
-    this.child,
+    required Animation<T> listenable,
+    required this.activeColor,
+    required this.inactiveColor,
+    required this.child,
   }) : super(listenable: listenable);
   final Color activeColor;
   final Color inactiveColor;
@@ -125,11 +129,11 @@ class _IconToggleable<T> extends AnimatedWidget {
 
 class _IconPainter extends CustomPainter {
   _IconPainter({
-    @required this.position,
-    this.activeColor,
-    this.inactiveColor,
+    required this.position,
+    required this.activeColor,
+    required this.inactiveColor,
   });
-  final Animation<double> position;
+  final position;
   final Color activeColor;
   final Color inactiveColor;
 
@@ -138,7 +142,7 @@ class _IconPainter extends CustomPainter {
   @override
   void paint(Canvas canvas, Size size) {
     final Paint paint = Paint()
-      ..color = Color.lerp(inactiveColor, activeColor, _value)
+      ..color = Color.lerp(inactiveColor, activeColor, _value)!
           .withOpacity(math.min(_value, 0.15))
       ..style = PaintingStyle.fill
       ..strokeWidth = 2.0;
